+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/), and this project
 adheres to [Semantic Versioning](https://semver.org/).
 
+## [Unreleased]
+
+[unreleased]: https://github.com/rogdham/bigxml/compare/v0.3.1...HEAD
+
+### :bug: Fixes
+
+- Typing: use `BinaryIO` instead of `IO[bytes]`
+
 ## [0.3.1] - 2021-12-26
 
 [0.3.1]: https://github.com/rogdham/python-xz/releases/tag/v0.3.1